feat: Add geolocation-based job filtering with timezone support
- Add optional timezone field to job posting form (IANA timezone format)
- Implement backend timezone filtering logic (±3 hour overlap)
- Add timezone filter to job browse page sidebar
- Support browser geolocation and manual timezone selection
- Jobs without timezone are visible to all users
- Install date-fns-tz for timezone calculations

Closes #96
