+++ backend/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@stellar/stellar-sdk": "^12.0.0",
     "cors": "^2.8.5",
+    "date-fns-tz": "^3.2.0",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "express-rate-limit": "^7.3.1",
