+++ backend/src/services/jobService.js
@@ -5,6 +5,7 @@
 "use strict";
 
 import { query } from "../db/pool";
+import { getTimezoneOffset } from "date-fns-tz";
 
 // ─── constants ───────────────────────────────────────────────────────────────
 
@@ -30,6 +31,37 @@ function validatePublicKey(key) {
     const e = new Error("Invalid Stellar public key");
     e.status = 400;
     throw e;
+  }
+}
+
+/**
+ * Check if a job's timezone is compatible with the user's timezone.
+ * Compatible if the time difference is within ±3 hours.
+ * 
+ * @param {string} jobTimezone - IANA timezone string of the job (e.g., "America/New_York")
+ * @param {string} userTimezone - IANA timezone string of the user (e.g., "Europe/London")
+ * @returns {boolean} true if timezones are compatible or if job has no timezone restriction
+ */
+function isTimezoneCompatible(jobTimezone, userTimezone) {
+  // Jobs without timezone restriction are visible to all users
+  if (!jobTimezone) return true;
+  // If no user timezone provided, show all jobs
+  if (!userTimezone) return true;
+
+  try {
+    const now = new Date();
+    // Get UTC offset in milliseconds for both timezones
+    const userOffset = getTimezoneOffset(userTimezone, now);
+    const jobOffset = getTimezoneOffset(jobTimezone, now);
+    
+    // Calculate the absolute difference in hours
+    const diffHours = Math.abs(userOffset - jobOffset) / (1000 * 60 * 60);
+    
+    // Return true if within ±3 hour range
+    return diffHours <= 3;
+  } catch (err) {
+    // If timezone parsing fails, show the job (fail-safe)
+    return true;
   }
 }
 
@@ -127,7 +159,7 @@ function decodeCursor(cursor) {
   }
 }
 
-async function listJobs({ category, status = "open", limit = 50, search, cursor } = {}) {
+async function listJobs({ category, status = "open", limit = 50, search, cursor, timezone } = {}) {
   const conditions = [];
   const params     = [];
 
@@ -172,8 +204,15 @@ async function listJobs({ category, status = "open", limit = 50, search, cursor 
   const currentRows = hasMore ? rows.slice(0, safeLimit) : rows;
   const nextCursor = hasMore ? encodeCursor(currentRows[currentRows.length - 1]) : null;
 
+  // Apply timezone filtering on the results
+  // This is done after fetching to avoid complex SQL timezone calculations
+  let filteredJobs = currentRows.map(rowToJob);
+  if (timezone) {
+    filteredJobs = filteredJobs.filter(job => isTimezoneCompatible(job.timezone, timezone));
+  }
+
   return {
-    jobs: currentRows.map(rowToJob),
+    jobs: filteredJobs,
     nextCursor,
   };
 }
